+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- implementations of `Add` and `AddAssign` to `manyhow::Error`/`manyhow::ErrorMessage`
+
 ## [0.11.2] - 2024-07-20
 ### Fixed
 - adjusted spans for `#[manyhow]` on `use` items to make go-to-definition work better.